+++ tests/testthat/test-allelic_series.R
@@ -193,8 +193,10 @@ test_that("Overall check of omnibus test.", {
     include_orig_skato_all = FALSE,
     include_orig_skato_ptv = FALSE
   )
-  expect_equal(length(p_omni), 3 + 8)
-  expect_equal(as.numeric(p_omni["p_omni"]), 0.0, tolerance = 0.005)
+  pvals <- p_omni@Pvals
+  p_omni <- as.numeric(pvals$pval[pvals$test == "omni"])
+  expect_equal(length(pvals$pval), 8)
+  expect_equal(p_omni, 0.0, tolerance = 0.005)
 
   # With SKAT-O all.
   p_omni <- COAST(
@@ -204,8 +206,10 @@ test_that("Overall check of omnibus test.", {
     include_orig_skato_all = TRUE,
     include_orig_skato_ptv = FALSE
   )
-  expect_equal(length(p_omni), 3 + 8 + 1)
-  expect_equal(as.numeric(p_omni["p_omni"]), 0.0, tolerance = 0.005)
+  pvals <- p_omni@Pvals
+  p_omni <- as.numeric(pvals$pval[pvals$test == "omni"])
+  expect_equal(length(pvals$pval), 8 + 1)
+  expect_equal(p_omni, 0.0, tolerance = 0.005)
 
   # With SKAT-O PTV.
   p_omni <- COAST(
@@ -215,8 +219,10 @@ test_that("Overall check of omnibus test.", {
     include_orig_skato_all = FALSE,
     include_orig_skato_ptv = TRUE
   )
-  expect_equal(length(p_omni), 3 + 8 + 1)
-  expect_equal(as.numeric(p_omni["p_omni"]), 0.0, tolerance = 0.005)
+  pvals <- p_omni@Pvals
+  p_omni <- as.numeric(pvals$pval[pvals$test == "omni"])
+  expect_equal(length(pvals$pval), 8 + 1)
+  expect_equal(p_omni, 0.0, tolerance = 0.005)
 
   # With both SKAT-O all and SKAT-O PTV.
   p_omni <- COAST(
@@ -226,8 +232,10 @@ test_that("Overall check of omnibus test.", {
     include_orig_skato_all = TRUE,
     include_orig_skato_ptv = TRUE
   )
-  expect_equal(length(p_omni), 3 + 8 + 2)
-  expect_equal(as.numeric(p_omni["p_omni"]), 0.0, tolerance = 0.005)
+  pvals <- p_omni@Pvals
+  p_omni <- as.numeric(pvals$pval[pvals$test == "omni"])
+  expect_equal(length(pvals$pval), 8 + 2)
+  expect_equal(p_omni, 0.0, tolerance = 0.005)
 
 })