+++ NEWS.md
@@ -1,3 +1,8 @@
+
+## Version 0.6.0
+* Replaced the `CountAlleles` function with a similar (but faster) `Counts` function that returns counts the total number of alleles, variants, and carriers by variant class.
+* Updated the formatting of the results output by the main `COAST` function.
+
 ## Version 0.5.0
 
 * Added option (`min_mac`) to filter the variant set to only include those variants having at least a minimum minor allele count (10 is recommended).